backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
alembic==1.13.1
SQLAlchemy==2.0.30
pydantic==2.7.3
requests==2.32.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
